+++ test/src/providers.spec.js
@@ -1,4 +1,11 @@
-import { behaviorsModel, behavior, fixturesModel, fixture } from "../../src/providers";
+import {
+  behaviorsModel,
+  behavior,
+  fixturesModel,
+  fixture,
+  alertsModel,
+  alert,
+} from "../../src/providers";
 
 describe("behaviorsModel findByName custom query", () => {
   it("should return name urlParam", () => {
@@ -29,5 +36,21 @@ describe("fixturesModel findById custom query", () => {
 describe("fixture alias", () => {
   it("should return queried fixturesModel", () => {
     expect(fixture("foo")).toEqual(fixturesModel.queries.byId("foo"));
+  });
+});
+
+describe("alertsModel findById custom query", () => {
+  it("should return name urlParam", () => {
+    expect(alertsModel.queryMethods.byId("foo")).toEqual({
+      urlParams: {
+        id: "foo",
+      },
+    });
+  });
+});
+
+describe("alert alias", () => {
+  it("should return queried alertsModel", () => {
+    expect(alert("foo")).toEqual(alertsModel.queries.byId("foo"));
   });
 });